% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_sampler_chain_n}
\alias{llama_sampler_chain_n}
\title{Number of samplers in a chain}
\usage{
llama_sampler_chain_n(chain)
}
\arguments{
\item{chain}{A sampler chain.}
}
\value{
Integer count of samplers currently in the chain.
}
\description{
Number of samplers in a chain
}
\seealso{
[llama_sampler_chain_get]
}
